LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-try-001-expected.txt


FAIL .target 1 assert_equals:
<div class="target" data-offset-x="150" data-offset-y="25" data-expected-width="40" data-expected-height="15"></div>
offsetLeft expected 150 but got 5
FAIL .target 2 assert_equals:
<div class="target" data-offset-x="5" data-offset-y="25" data-expected-width="40" data-expected-height="15"></div>
offsetTop expected 25 but got 5
FAIL .target 3 assert_equals:
<div class="target" data-offset-x="50" data-offset-y="35" data-expected-width="40" data-expected-height="15"></div>
offsetLeft expected 50 but got 5
FAIL .target 4 assert_equals:
<div class="target" data-offset-x="50" data-offset-y="20" data-expected-width="40" data-expected-height="15"></div>
offsetLeft expected 50 but got 5
FAIL .target 5 assert_equals:
<div class="target" data-offset-x="150" data-offset-y="25" data-expected-width="35" data-expected-height="40"></div>
width expected 35 but got 40
FAIL .target 6 assert_equals:
<div class="target" data-offset-x="150" data-offset-y="25" data-expected-width="40" data-expected-height="15"></div>
offsetLeft expected 150 but got 5
